requirements.txt
requests
networkx
matplotlib
flask
plotly
dash
dash-cytoscape
python-dotenv
cryptography
